requirements.txt
langchain
langchain_community
pandas
pymongo
python-dotenv
streamlit
faiss-cpu
google-genai
